+++ Magic/src/main/java/com/elmakers/mine/bukkit/world/populator/builtin/EndPortalPopulator.java
@@ -16,18 +16,16 @@ import com.elmakers.mine.bukkit.world.populator.BaseBlockPopulator;
 public class EndPortalPopulator extends BaseBlockPopulator {
     private int minPortalWidth = 3;
     private int maxPortalWidth = 3;
-    private int minTunnelHeight = 0;
-    private int maxTunnelHeight = 0;
-    private boolean gateway = false;
+    private int minGatewayHeight = 0;
+    private int maxGatewayHeight = 0;
     private String targetWorld;
 
     @Override
     public boolean onLoad(ConfigurationSection config) {
         minPortalWidth = config.getInt("min_portal_width", minPortalWidth);
         maxPortalWidth = config.getInt("max_portal_width", maxPortalWidth);
-        minTunnelHeight = config.getInt("min_tunnel_height", minPortalWidth);
-        maxTunnelHeight = config.getInt("max_tunnel_height", maxPortalWidth);
-        gateway = config.getBoolean("gateway", gateway);
+        minGatewayHeight = config.getInt("min_gateway_height", minPortalWidth);
+        maxGatewayHeight = config.getInt("max_gateway_height", maxPortalWidth);
         targetWorld = config.getString("target_world", targetWorld);
         return true;
     }
@@ -47,35 +45,23 @@ public class EndPortalPopulator extends BaseBlockPopulator {
         final int chunkGlobalZ = chunkZ << 4;
         final int floorLevel = world.getGroundLevel();
         final int portalWidth = RandomUtils.range(random, minPortalWidth, maxPortalWidth);
-        final int tunnelHeight = RandomUtils.range(random, minTunnelHeight, maxTunnelHeight);
+        final int gatewayHeight = RandomUtils.range(random, minGatewayHeight, maxGatewayHeight);
         final int portalLeft = 8 - (int)Math.ceil((double)portalWidth / 2);
         final int portalRight = 8 + (int)Math.floor((double)portalWidth / 2);
 
-        final int buffer = region.getBuffer();
-        final int minX = chunkGlobalX - buffer;
-        final int maxX = chunkGlobalX + 16 + buffer;
-        final int minZ = chunkGlobalZ - buffer;
-        final int maxZ = chunkGlobalZ + 16 + buffer;
         final int minExitX = chunkGlobalX + portalLeft;
         final int maxExitX = chunkGlobalX + portalRight;
         final int minExitZ = chunkGlobalZ + portalLeft;
         final int maxExitZ = chunkGlobalZ + portalRight;
-        final int maxExitY = floorLevel + tunnelHeight;
-        final BlockData exitBlock = gateway ? getGatewayBlock() : null;
+        final int maxExitY = floorLevel + gatewayHeight;
+        final BlockData exitBlock = gatewayHeight > 0 ? getGatewayBlock() : null;
 
-        for (int x = minX; x < maxX; x++) {
-            for (int z = minZ; z < maxZ; z++) {
-                for (int y = floorLevel + 1; y <= maxExitY; y++) {
-                    if (exitBlock != null && x >= minExitX && x <= maxExitX && z >= minExitZ && z <= maxExitZ) {
-                        region.setBlockData(x, y, z, exitBlock);
-                    } else {
-                        region.setType(x, y, z, Material.AIR);
-                    }
+        for (int x = minExitX; x < maxExitX; x++) {
+            for (int z = minExitZ; z < maxExitZ; z++) {
+                for (int y = floorLevel + 1; y <= maxExitY && exitBlock != null; y++) {
+                    region.setBlockData(x, y, z, exitBlock);
                 }
-
-                if (x >= minExitX && x <= maxExitX && z >= minExitZ && z <= maxExitZ) {
-                    region.setType(x, floorLevel, z, Material.END_PORTAL);
-                }
+                region.setType(x, floorLevel, z, Material.END_PORTAL);
             }
         }
     }